+++ biceps/src/main/java/org/somda/sdc/biceps/consumer/access/RemoteMdibAccessImpl.java
@@ -1,5 +1,6 @@
 package org.somda.sdc.biceps.consumer.access;
 
+import com.google.inject.Injector;
 import com.google.inject.assistedinject.AssistedInject;
 import com.google.inject.name.Named;
 import org.apache.logging.log4j.LogManager;
@@ -14,13 +15,14 @@ import org.somda.sdc.biceps.common.access.WriteStateResult;
 import org.somda.sdc.biceps.common.access.factory.ReadTransactionFactory;
 import org.somda.sdc.biceps.common.access.helper.WriteUtil;
 import org.somda.sdc.biceps.common.event.Distributor;
-import org.somda.sdc.biceps.common.preprocessing.DescriptorChildRemover;
+import org.somda.sdc.biceps.common.preprocessing.PreprocessingUtil;
+import org.somda.sdc.biceps.common.storage.DescriptionPreprocessingSegment;
 import org.somda.sdc.biceps.common.storage.MdibStorage;
 import org.somda.sdc.biceps.common.storage.MdibStoragePreprocessingChain;
 import org.somda.sdc.biceps.common.storage.PreprocessingException;
+import org.somda.sdc.biceps.common.storage.StatePreprocessingSegment;
 import org.somda.sdc.biceps.common.storage.factory.MdibStorageFactory;
 import org.somda.sdc.biceps.common.storage.factory.MdibStoragePreprocessingChainFactory;
-import org.somda.sdc.biceps.consumer.preprocessing.VersionDuplicateHandler;
 import org.somda.sdc.biceps.model.participant.AbstractContextState;
 import org.somda.sdc.biceps.model.participant.AbstractDescriptor;
 import org.somda.sdc.biceps.model.participant.AbstractState;
@@ -30,7 +32,6 @@ import org.somda.sdc.common.logging.InstanceLogger;
 
 import javax.annotation.Nullable;
 import java.math.BigInteger;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
@@ -57,19 +58,28 @@ public class RemoteMdibAccessImpl implements RemoteMdibAccess {
                          MdibStorageFactory mdibStorageFactory,
                          ReentrantReadWriteLock readWriteLock,
                          ReadTransactionFactory readTransactionFactory,
-                         VersionDuplicateHandler versionDuplicateHandler,
-                         DescriptorChildRemover descriptorChildRemover,
-                         @Named(CommonConfig.INSTANCE_IDENTIFIER) String frameworkIdentifier) {
+                         @Named(CommonConfig.INSTANCE_IDENTIFIER) String frameworkIdentifier,
+                         @Named(org.somda.sdc.biceps.common.CommonConfig.CONSUMER_STATE_PREPROCESSING_SEGMENTS)
+                                 List<Class<? extends StatePreprocessingSegment>> stateSegmentClasses,
+                         @Named(org.somda.sdc.biceps.common.CommonConfig.CONSUMER_DESCRIPTION_PREPROCESSING_SEGMENTS)
+                                 List<Class<? extends DescriptionPreprocessingSegment>> descriptionSegmentClasses,
+                         Injector injector) {
         this.instanceLogger = InstanceLogger.wrapLogger(LOG, frameworkIdentifier);
         this.eventDistributor = eventDistributor;
         this.mdibStorage = mdibStorageFactory.createMdibStorage();
         this.readWriteLock = readWriteLock;
         this.readTransactionFactory = readTransactionFactory;
 
+        var descriptionPreprocessingSegments =
+                PreprocessingUtil.getDescriptionPreprocessingSegments(descriptionSegmentClasses, injector);
+
+        var statePreprocessingSegments = PreprocessingUtil.getStatePreprocessingSegments(
+                stateSegmentClasses, descriptionPreprocessingSegments, injector);
+
         this.localMdibAccessPreprocessing = chainFactory.createMdibStoragePreprocessingChain(
                 mdibStorage,
-                Arrays.asList(descriptorChildRemover),
-                Arrays.asList(versionDuplicateHandler));
+                descriptionPreprocessingSegments,
+                statePreprocessingSegments);
 
         this.writeUtil = new WriteUtil(
                 instanceLogger, eventDistributor,