+++ biceps/src/test/java/org/somda/sdc/biceps/common/storage/MdibStorageImplTest.java
@@ -1,6 +1,7 @@
 package org.somda.sdc.biceps.common.storage;
 
 import com.google.inject.Injector;
+import com.google.inject.TypeLiteral;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -10,13 +11,39 @@ import org.somda.sdc.biceps.common.CommonConfig;
 import org.somda.sdc.biceps.common.MdibDescriptionModification;
 import org.somda.sdc.biceps.common.MdibDescriptionModifications;
 import org.somda.sdc.biceps.common.MdibStateModifications;
+import org.somda.sdc.biceps.common.preprocessing.DescriptorChildRemover;
 import org.somda.sdc.biceps.common.storage.factory.MdibStorageFactory;
-import org.somda.sdc.biceps.model.participant.*;
+import org.somda.sdc.biceps.consumer.preprocessing.DuplicateContextStateHandleHandler;
+import org.somda.sdc.biceps.consumer.preprocessing.VersionDuplicateHandler;
+import org.somda.sdc.biceps.model.participant.AbstractContextState;
+import org.somda.sdc.biceps.model.participant.AlertConditionDescriptor;
+import org.somda.sdc.biceps.model.participant.AlertConditionState;
+import org.somda.sdc.biceps.model.participant.AlertSystemDescriptor;
+import org.somda.sdc.biceps.model.participant.AlertSystemState;
+import org.somda.sdc.biceps.model.participant.ChannelDescriptor;
+import org.somda.sdc.biceps.model.participant.ChannelState;
+import org.somda.sdc.biceps.model.participant.ContextAssociation;
+import org.somda.sdc.biceps.model.participant.EnsembleContextDescriptor;
+import org.somda.sdc.biceps.model.participant.EnsembleContextState;
+import org.somda.sdc.biceps.model.participant.LocationContextDescriptor;
+import org.somda.sdc.biceps.model.participant.LocationContextState;
+import org.somda.sdc.biceps.model.participant.MdibVersion;
+import org.somda.sdc.biceps.model.participant.MdsDescriptor;
+import org.somda.sdc.biceps.model.participant.MdsState;
+import org.somda.sdc.biceps.model.participant.PatientContextDescriptor;
+import org.somda.sdc.biceps.model.participant.PatientContextState;
+import org.somda.sdc.biceps.model.participant.SystemContextDescriptor;
+import org.somda.sdc.biceps.model.participant.SystemContextState;
+import org.somda.sdc.biceps.model.participant.VmdDescriptor;
+import org.somda.sdc.biceps.model.participant.VmdState;
+import org.somda.sdc.biceps.provider.preprocessing.DuplicateChecker;
+import org.somda.sdc.biceps.provider.preprocessing.HandleReferenceHandler;
+import org.somda.sdc.biceps.provider.preprocessing.TypeConsistencyChecker;
+import org.somda.sdc.biceps.provider.preprocessing.VersionHandler;
 import org.somda.sdc.biceps.testutil.Handles;
 import org.somda.sdc.biceps.testutil.MockModelFactory;
 import org.somda.sdc.common.guice.AbstractConfigurationModule;
 import test.org.somda.common.LoggingTestWatcher;
-import test.org.somda.common.TestLogging;
 
 import java.math.BigInteger;
 import java.util.Arrays;
@@ -24,7 +51,11 @@ import java.util.List;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 
 @ExtendWith(LoggingTestWatcher.class)
@@ -465,6 +496,26 @@ public class MdibStorageImplTest {
                 bind(CommonConfig.COPY_MDIB_OUTPUT, Boolean.class, true);
                 bind(CommonConfig.STORE_NOT_ASSOCIATED_CONTEXT_STATES, Boolean.class, true);
                 bind(CommonConfig.ALLOW_STATES_WITHOUT_DESCRIPTORS, Boolean.class, true);
+                bind(CommonConfig.CONSUMER_STATE_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends StatePreprocessingSegment>>>() {
+                        },
+                        List.of(DuplicateContextStateHandleHandler.class, VersionDuplicateHandler.class));
+
+                bind(CommonConfig.CONSUMER_DESCRIPTION_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends DescriptionPreprocessingSegment>>>() {
+                        },
+                        List.of(DescriptorChildRemover.class));
+
+                bind(CommonConfig.PROVIDER_STATE_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends StatePreprocessingSegment>>>() {
+                        },
+                        List.of(DuplicateContextStateHandleHandler.class, VersionHandler.class));
+
+                bind(CommonConfig.PROVIDER_DESCRIPTION_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends DescriptionPreprocessingSegment>>>() {
+                        },
+                        List.of(DuplicateChecker.class, TypeConsistencyChecker.class, VersionHandler.class,
+                                HandleReferenceHandler.class, DescriptorChildRemover.class));
             }
         }).getInjector();
         return injector.getInstance(MdibStorageFactory.class).createMdibStorage();
@@ -616,6 +667,26 @@ public class MdibStorageImplTest {
                 bind(CommonConfig.COPY_MDIB_OUTPUT, Boolean.class, true);
                 bind(CommonConfig.STORE_NOT_ASSOCIATED_CONTEXT_STATES, Boolean.class, true);
                 bind(CommonConfig.ALLOW_STATES_WITHOUT_DESCRIPTORS, Boolean.class, false);
+                bind(CommonConfig.CONSUMER_STATE_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends StatePreprocessingSegment>>>() {
+                        },
+                        List.of(DuplicateContextStateHandleHandler.class, VersionDuplicateHandler.class));
+
+                bind(CommonConfig.CONSUMER_DESCRIPTION_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends DescriptionPreprocessingSegment>>>() {
+                        },
+                        List.of(DescriptorChildRemover.class));
+
+                bind(CommonConfig.PROVIDER_STATE_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends StatePreprocessingSegment>>>() {
+                        },
+                        List.of(DuplicateContextStateHandleHandler.class, VersionHandler.class));
+
+                bind(CommonConfig.PROVIDER_DESCRIPTION_PREPROCESSING_SEGMENTS,
+                        new TypeLiteral<List<Class<? extends DescriptionPreprocessingSegment>>>() {
+                        },
+                        List.of(DuplicateChecker.class, TypeConsistencyChecker.class, VersionHandler.class,
+                                HandleReferenceHandler.class, DescriptorChildRemover.class));
             }
         }).getInjector();
         var localMdibStorage = injector.getInstance(MdibStorageFactory.class).createMdibStorage();